+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Fixes:
 * [Issue #70](https://github.com/apptentive/apptentive-ios/issues/70) unreadCount incremented when typing a message
 * IOS-310 unreadMessageCount incremented when typing a message
 * IOS-341 Investigate loading of initial config file at app launch.
+* IOS-355 Strip whitespace from Survey text response
 * IOS-356 Can't see send button in Message Center with white tintColor
 * IOS-357 Add support for hidden text/file messages